Move remaining logic to commandDetails
